Refactor model and service layers: rename packages, update entity fields, and implement mappers for Location and Transportation

+++ src/main/java/com/fatihbozik/aviationroutefinder/repository/LocationRepository.java
@@ -1,10 +1,10 @@
 package com.fatihbozik.aviationroutefinder.repository;
 
-import com.fatihbozik.aviationroutefinder.model.LocationEntity;
+import com.fatihbozik.aviationroutefinder.persistence.LocationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LocationRepository extends JpaRepository<LocationEntity, Integer> {
+public interface LocationRepository extends JpaRepository<LocationEntity, Long> {
 
 }
